ignore errors when cleaning up resources
When doing an uninstall it's not really an error if we try to remove
something that has already been removed. Add a few checks to cover this
scenario.

There's an upstream patch proposed which handles this a bit differently
which we may want to pick up instead, it basically just ignores all
errors on uninstall rather than being selective.
https://github.com/intel/CPU-Manager-for-Kubernetes/pull/239

Signed-off-by: Chris Friesen <chris.friesen@windriver.com>