improve the uninstall command in cmk.py
The upstream "cmk uninstall" command will both clean up the node it
runs on as well as try to uninstall the global webhook.  This is
not ideal if we just want to do a helm upgrade to modify the CPU
allocation.  The command will also try to clean up a bunch of pods
which are no longer applicable when using a helm chart.

Accordingly the "uninstall" command is modified to no longer clean
up the pods and to leave the webhook resources alone.  A new
"uninstall-webhook" command is created to clean up the webhook
resources.

Signed-off-by: Chris Friesen <chris.friesen@windriver.com>