Default to allocating all available isolated CPUs as exclusive
By default we want to allocate from the "isolated" CPUs a single
"shared" CPU and then use all the rest of the "isolated" CPUs as
"exclusive" CPUs.

To accomplish this we modify the code to treat a core count of
-1 as meaning "all available cores".  We then allocate the "shared"
cores first and the "exclusive" cores second, and we change the
default number of "exclusive" cores to -1.

Signed-off-by: Chris Friesen <chris.friesen@windriver.com>